fix:解决token拦截器获取post请求的token失败问题，实现点赞接口

+++ daos/video_dao.go
@@ -12,7 +12,7 @@ var (
 )
 
 type VideoDao interface {
-	FindById(id int) (*models.Video, error)
+	FindById(id int64) (*models.Video, error)
 }
 
 type VideoDaoImpl struct {
@@ -28,7 +28,7 @@ func GetVideoDao() VideoDao {
 	return videoDao
 }
 
-func (v *VideoDaoImpl) FindById(id int) (*models.Video, error) {
+func (v *VideoDaoImpl) FindById(id int64) (*models.Video, error) {
 	var video models.Video
 	err := v.db.Debug().Select("id", "user_id", "play_url", "cover_url", "comment_count", "favorite_count", "title").
 		Where("id = ?", id).Take(&video).Error
